move parallel-for stuff into cpp
